replace old context/kernel with new ones, make fields readonly, port tests to new api
