+++ expander_compiler/Cargo.toml
@@ -57,6 +57,6 @@ path = "bin/trivial_circuit.rs"
 name = "expander_serve"
 path = "src/zkcuda/proving_system/expander_gkr_parallelized/server_bin.rs"
 
-#[[bin]]
-#name = "matmul_bench"
-#path = "bin/matmul_bench.rs"
+[[bin]]
+name = "matmul_bench"
+path = "bin/matmul_bench.rs"
