+++ expander_compiler/src/zkcuda/mod.rs
@@ -1,8 +1,5 @@
 pub mod context;
-pub mod context2;
 pub mod kernel;
-pub mod kernel2;
-pub mod proof;
 pub mod proving_system;
 pub mod shape;
 pub mod vec_shaped;
