--- expander_compiler/src/zkcuda/proof.rs
@@ -1,20 +1,0 @@
-use serdes::ExpSerde;
-
-use super::kernel::Kernel;
-
-use crate::circuit::config::Config;
-
-#[derive(Clone, ExpSerde)]
-pub struct ProofTemplate {
-    pub kernel_id: usize,
-    pub commitment_indices: Vec<usize>,
-    pub parallel_count: usize,
-    pub is_broadcast: Vec<bool>,
-}
-
-#[derive(Default, ExpSerde)]
-pub struct ComputationGraph<C: Config> {
-    pub kernels: Vec<Kernel<C>>,
-    pub commitments_lens: Vec<usize>,
-    pub proof_templates: Vec<ProofTemplate>,
-}
