+++ expander_compiler/src/zkcuda/proving_system/dummy.rs
@@ -3,8 +3,7 @@ use serdes::ExpSerde;
 
 use crate::circuit::config::{Config, SIMDField};
 use crate::utils::misc::next_power_of_two;
-use crate::zkcuda::context::DeviceMemory;
-use crate::zkcuda::proof::ComputationGraph;
+use crate::zkcuda::context::ComputationGraph;
 use crate::zkcuda::proving_system::{CombinedProof, KernelWiseProvingSystem, ProvingSystem};
 
 use super::super::kernel::Kernel;
@@ -45,12 +44,10 @@ impl<C: Config> KernelWiseProvingSystem<C> for DummyProvingSystem<C> {
     type Commitment = DummyCommitment<C>;
     type CommitmentExtraInfo = ();
 
-    fn setup(
-        computation_graph: &crate::zkcuda::proof::ComputationGraph<C>,
-    ) -> (Self::ProverSetup, Self::VerifierSetup) {
+    fn setup(computation_graph: &ComputationGraph<C>) -> (Self::ProverSetup, Self::VerifierSetup) {
         // let _ = computation_graph;
         computation_graph
-            .commitments_lens
+            .commitments_lens()
             .iter()
             .for_each(|&x| println!("Setup length {x}"));
 
@@ -86,14 +83,14 @@ impl<C: Config> KernelWiseProvingSystem<C> for DummyProvingSystem<C> {
         let mut res = vec![];
         for i in 0..parallel_count {
             let lc_input = prepare_inputs(
-                &kernel.layered_circuit,
-                &kernel.layered_circuit_input,
+                &kernel.layered_circuit(),
+                &kernel.layered_circuit_input(),
                 commitments_values,
                 is_broadcast,
                 i,
             );
             let (_, cond) = kernel
-                .layered_circuit
+                .layered_circuit()
                 .eval_with_public_inputs_simd(lc_input, &[]);
             for x in cond.iter() {
                 if !*x {
@@ -117,14 +114,14 @@ impl<C: Config> KernelWiseProvingSystem<C> for DummyProvingSystem<C> {
         check_inputs(kernel, &values, parallel_count, is_broadcast);
         for i in 0..parallel_count {
             let lc_input = prepare_inputs(
-                &kernel.layered_circuit,
-                &kernel.layered_circuit_input,
+                &kernel.layered_circuit(),
+                &kernel.layered_circuit_input(),
                 &values,
                 is_broadcast,
                 i,
             );
             let (_, cond) = kernel
-                .layered_circuit
+                .layered_circuit()
                 .eval_with_public_inputs_simd(lc_input, &[]);
             if cond != proof.cond[i] {
                 return false;
@@ -153,21 +150,21 @@ impl<C: Config> ProvingSystem<C> for DummyProvingSystem<C> {
     fn prove(
         prover_setup: &Self::ProverSetup,
         computation_graph: &ComputationGraph<C>,
-        device_memories: &[DeviceMemory<C>],
+        device_memories: &[Vec<SIMDField<C>>],
     ) -> Self::Proof {
         let commitments = computation_graph
-            .proof_templates
+            .proof_templates()
             .iter()
             .map(|template| {
                 template
-                    .commitment_indices
+                    .commitment_indices()
                     .iter()
-                    .zip(template.is_broadcast.iter())
+                    .zip(template.is_broadcast().iter())
                     .map(|(x, is_broadcast)| {
                         <Self as KernelWiseProvingSystem<C>>::commit(
                             prover_setup,
-                            &device_memories[*x].values,
-                            next_power_of_two(template.parallel_count),
+                            &device_memories[*x],
+                            next_power_of_two(template.parallel_count()),
                             *is_broadcast,
                         )
                     })
@@ -176,23 +173,23 @@ impl<C: Config> ProvingSystem<C> for DummyProvingSystem<C> {
             .collect::<Vec<_>>();
 
         let proofs: Vec<<Self as KernelWiseProvingSystem<C>>::Proof> = computation_graph
-            .proof_templates
+            .proof_templates()
             .iter()
             .zip(commitments.iter())
             .map(|(template, commitments_kernel)| {
                 <Self as KernelWiseProvingSystem<C>>::prove_kernel(
                     prover_setup,
-                    template.kernel_id,
-                    &computation_graph.kernels[template.kernel_id],
+                    template.kernel_id(),
+                    &computation_graph.kernels()[template.kernel_id()],
                     &commitments_kernel.0,
                     &commitments_kernel.1,
                     &template
-                        .commitment_indices
+                        .commitment_indices()
                         .iter()
-                        .map(|x| &device_memories[*x].values[..])
+                        .map(|x| &device_memories[*x][..])
                         .collect::<Vec<_>>(),
-                    next_power_of_two(template.parallel_count),
-                    &template.is_broadcast,
+                    next_power_of_two(template.parallel_count()),
+                    &template.is_broadcast(),
                 )
             })
             .collect::<Vec<_>>();
@@ -211,17 +208,17 @@ impl<C: Config> ProvingSystem<C> for DummyProvingSystem<C> {
         let verified = proof
             .proofs
             .par_iter()
-            .zip(computation_graph.proof_templates.par_iter())
+            .zip(computation_graph.proof_templates().par_iter())
             .zip(proof.commitments.par_iter())
             .map(|((proof, template), commitments_kernel)| {
                 <Self as KernelWiseProvingSystem<C>>::verify_kernel(
                     verifier_setup,
-                    template.kernel_id,
-                    &computation_graph.kernels[template.kernel_id],
+                    template.kernel_id(),
+                    &computation_graph.kernels()[template.kernel_id()],
                     proof,
                     commitments_kernel,
-                    next_power_of_two(template.parallel_count),
-                    &template.is_broadcast,
+                    next_power_of_two(template.parallel_count()),
+                    &template.is_broadcast(),
                 )
             })
             .collect::<Vec<_>>();
