+++ expander_compiler/src/zkcuda/proving_system/expander_gkr_parallelized/interface.rs
@@ -2,6 +2,7 @@ use std::fs;
 use std::io::{Cursor, Read};
 
 use crate::circuit::config::Config;
+use crate::frontend::SIMDField;
 use crate::utils::misc::next_power_of_two;
 use crate::zkcuda::kernel::Kernel;
 use crate::zkcuda::proving_system::expander_gkr_parallelized::client::{
@@ -53,7 +54,7 @@ where
     ) -> bool {
         let timer = Timer::new("verify", true);
 
-        let mut expander_circuit = kernel.layered_circuit.export_to_expander().flatten::<C>();
+        let mut expander_circuit = kernel.layered_circuit().export_to_expander().flatten::<C>();
         expander_circuit.pre_process_gkr::<C>();
 
         let mut transcript = C::TranscriptConfig::new();
@@ -117,7 +118,7 @@ where
         assert_eq!(1 << challenge.r_mpi.len(), parallel_count);
         let mut target_y = <C::FieldConfig as FieldEngine>::ChallengeField::ZERO;
         for ((input, commitment), ib) in kernel
-            .layered_circuit_input
+            .layered_circuit_input()
             .iter()
             .zip(commitments)
             .zip(is_broadcast)
@@ -207,7 +208,7 @@ where
     type Proof = CombinedProof<ECCConfig, ExpanderGKRProvingSystem<C>>;
 
     fn setup(
-        computation_graph: &crate::zkcuda::proof::ComputationGraph<ECCConfig>,
+        computation_graph: &crate::zkcuda::context::ComputationGraph<ECCConfig>,
     ) -> (Self::ProverSetup, Self::VerifierSetup) {
         let setup_timer = Timer::new("setup", true);
 
@@ -221,9 +222,9 @@ where
         fs::write(&setup_filename, bytes).expect("Failed to write computation graph to file");
 
         let max_parallel_count = computation_graph
-            .proof_templates
+            .proof_templates()
             .iter()
-            .map(|t| t.parallel_count)
+            .map(|t| t.parallel_count())
             .max()
             .unwrap_or(1);
 
@@ -247,16 +248,13 @@ where
 
     fn prove(
         _prover_setup: &Self::ProverSetup,
-        _computation_graph: &crate::zkcuda::proof::ComputationGraph<ECCConfig>,
-        device_memories: &[crate::zkcuda::context::DeviceMemory<ECCConfig>],
+        _computation_graph: &crate::zkcuda::context::ComputationGraph<ECCConfig>,
+        device_memories: &[Vec<SIMDField<ECCConfig>>],
     ) -> Self::Proof {
         let timer = Timer::new("prove", true);
 
         SharedMemoryEngine::write_witness_to_shared_memory::<C::FieldConfig>(
-            &device_memories
-                .iter()
-                .map(|m| &m.values[..])
-                .collect::<Vec<_>>(),
+            &device_memories.iter().map(|m| &m[..]).collect::<Vec<_>>(),
         );
         wait_async(request_prove());
 
@@ -266,23 +264,23 @@ where
 
     fn verify(
         verifier_setup: &Self::VerifierSetup,
-        computation_graph: &crate::zkcuda::proof::ComputationGraph<ECCConfig>,
+        computation_graph: &crate::zkcuda::context::ComputationGraph<ECCConfig>,
         proof: &Self::Proof,
     ) -> bool {
         let verified = proof
             .proofs
             .par_iter()
-            .zip(computation_graph.proof_templates.par_iter())
+            .zip(computation_graph.proof_templates().par_iter())
             .zip(proof.commitments.par_iter())
             .map(|((proof, template), commitments_kernel)| {
                 Self::verify_kernel(
                     verifier_setup,
-                    template.kernel_id,
-                    &computation_graph.kernels[template.kernel_id],
+                    template.kernel_id(),
+                    &computation_graph.kernels()[template.kernel_id()],
                     proof,
                     commitments_kernel,
-                    next_power_of_two(template.parallel_count),
-                    &template.is_broadcast,
+                    next_power_of_two(template.parallel_count()),
+                    &template.is_broadcast(),
                 )
             })
             .collect::<Vec<_>>();
